Extract CheckboxIcon
